Code/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.25.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torch-fidelity==0.3.0
torchvision==0.20.1
tqdm==4.64.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
